+++ lib/data/pagesLinks.ts
@@ -1,8 +1,9 @@
-import { PAGES_URL } from '@/lib/utils/constants/urls/pageUrls'
-import { TEXT } from '@/localization/english'
-
 import { DATA_TEST_IDS } from '@/__tests__/playwright/lib/utils/constants/ids/dataTestIds'
 import { ID } from '@/lib/utils/constants/ids/elementIds'
+import { PAGES_URL } from '@/lib/utils/constants/urls/pageUrls'
+
+import { TEXT } from '@/localization/english'
+
 import { PageLinkItem } from '@/lib/utils/interfaces/interfaces'
 
 export const pagesLinks: PageLinkItem[] = [
