+++ lib/data/projects/work/projects-overview/workFrontEnd.ts
@@ -1,10 +1,11 @@
 import { iconsSkillsProjectsWorkSmartsuppWeb } from '@/lib/data/skills/skillsProjectsWork'
 
 import { PROJECT_ID } from '@/lib/utils/constants/ids/projectIds'
+import { PAGES_URL } from '@/lib/utils/constants/urls/pageUrls'
+
 import { Project } from '@/lib/utils/interfaces/interfaces'
 
 // smartsupp web
-import { PAGES_URL } from '@/lib/utils/constants/urls/pageUrls'
 import smartsuppWeb1 from '@/public/images/webp/projects/work/smartsupp-web/smartsupp-web-1-light.webp'
 import smartsuppWeb2 from '@/public/images/webp/projects/work/smartsupp-web/smartsupp-web-2-light.webp'
 import smartsuppWeb3 from '@/public/images/webp/projects/work/smartsupp-web/smartsupp-web-3-light.webp'
