+++ lib/data/projects/work/projects-overview/workQA.ts
@@ -1,10 +1,11 @@
 import { iconsSkillsProjectsWorkGroupon } from '@/lib/data/skills/skillsProjectsWork'
 
 import { PROJECT_ID } from '@/lib/utils/constants/ids/projectIds'
+import { PAGES_URL } from '@/lib/utils/constants/urls/pageUrls'
+
 import { Project } from '@/lib/utils/interfaces/interfaces'
 
 // groupon
-import { PAGES_URL } from '@/lib/utils/constants/urls/pageUrls'
 import groupon1 from '@/public/images/webp/projects/work/groupon/groupon-1-light.webp'
 import groupon2 from '@/public/images/webp/projects/work/groupon/groupon-2-light.webp'
 import grouponThumb from '@/public/images/webp/projects/work/groupon/groupon-thumb-light-600.webp'
