+++ lib/data/projects/work/projects-overview/workWordpress.ts
@@ -1,10 +1,11 @@
 import { iconsSkillsProjectsWorkSmartsuppHelp } from '@/lib/data/skills/skillsProjectsWork'
 
 import { PROJECT_ID } from '@/lib/utils/constants/ids/projectIds'
+import { PAGES_URL } from '@/lib/utils/constants/urls/pageUrls'
+
 import { Project } from '@/lib/utils/interfaces/interfaces'
 
 // smartsupp help
-import { PAGES_URL } from '@/lib/utils/constants/urls/pageUrls'
 import smartsuppHelp1 from '@/public/images/webp/projects/work/smartsupp-help/smartsupp-help-1-light.webp'
 import smartsuppHelp2 from '@/public/images/webp/projects/work/smartsupp-help/smartsupp-help-2-light.webp'
 import smartsuppHelp3 from '@/public/images/webp/projects/work/smartsupp-help/smartsupp-help-3-light.webp'
