+++ lib/utils/helpers/breadcrumbs/getBreadcrumbsPersonal.ts
@@ -1,4 +1,5 @@
 import { URL_PERSONAL_PROJECTS } from '@/lib/utils/constants/urls/pageUrls'
+
 import { TEXT } from '@/localization/english'
 
 import { BreadCrumbsType } from '@/lib/utils/interfaces/types'
